pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = [
    {name = "Grigoriy Lavrov",email = "g.lavrov@vk.team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pylint (>=3.3.5,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "sphinx (>=8.2.3,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = true

